Creating Resolution26Screen style file
