+++ src/screens/legislation/Resolution26Screen.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import {
-  StyleSheet,
   Text,
   View,
   TouchableHighlight,
@@ -21,62 +20,7 @@ import {
   ANEXOS,
 } from '../../constants/legislation/resolutionNumber26';
 import { backHandlerPop } from '../../NavigationFunctions';
-
-const styles = StyleSheet.create({
-  title: {
-    textAlign: 'center',
-    fontSize: 22,
-    fontWeight: '300',
-    marginBottom: 20,
-  },
-  selectors: {
-    marginBottom: 10,
-    flexDirection: 'row',
-    justifyContent: 'center',
-  },
-  selector: {
-    backgroundColor: '#F5FCFF',
-    paddingBottom: 20,
-    paddingTop: 10,
-    paddingLeft: 10,
-    paddingRight: 10,
-  },
-  activeSelector: {
-    fontWeight: 'bold',
-  },
-  selectTitle: {
-    fontSize: 14,
-    fontWeight: '500',
-    padding: 10,
-    justifyContent: 'center',
-  },
-  field: {
-    backgroundColor: '#FAFAFA',
-    padding: 6,
-    borderWidth: 1,
-    borderRadius: 7,
-    borderColor: 'gray',
-    marginHorizontal: 15,
-    marginBottom: 15,
-    marginTop: 15,
-    justifyContent: 'flex-start',
-  },
-  buttonTexts: {
-    padding: 5,
-    color: 'white',
-    fontWeight: 'bold',
-    fontSize: 20,
-  },
-  fieldStyle: {
-    paddingHorizontal: 50,
-    paddingVertical: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#FF9500',
-    borderRadius: 8,
-    borderWidth: 1,
-  },
-});
+import styles from '../../Styles/LegislationStyles/Resolution26ScreenStyles'
 
 export default class Resolution26Screen extends Component {
   state = {
